Rename unit test for weather service
